server/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.0
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==24.3.0
Authlib==1.4.1
bcrypt==4.2.1
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
decorator==5.1.1
deprecation==2.1.0
dotenv==0.9.9
executing==2.1.0
Faker==14.2.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==3.1.0
Flask-OAuthlib==0.9.6
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
frozenlist==1.5.0
gotrue==2.11.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib-metadata==6.0.0
importlib-resources==5.10.0
iniconfig==2.0.0
ipdb==0.13.9
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.1.0
oauthlib==2.1.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pluggy==1.5.0
postgrest==0.19.3
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==7.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
realtime==2.3.0
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==1.1.0
resend==2.6.0
setuptools==70.3.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
stack-data==0.6.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
toml==0.10.2
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
websockets==14.2
Werkzeug==2.2.2
yarl==1.18.3
zipp==3.21.0
